[Added] Possibility to extend play and record service functions

+++ lib/python/Navigation.py
@@ -21,6 +21,8 @@ from Screens.InfoBarGenerics import streamrelay
 
 
 class Navigation:
+	playServiceExtensions = []
+	recordServiceExtensions = []
 	def __init__(self, nextRecordTimerAfterEventActionAuto=False, nextPowerManagerAfterEventActionAuto=False):
 		if NavigationInstance.instance is not None:
 			raise NavigationInstance.instance
@@ -36,6 +38,7 @@ class Navigation:
 		self.currentlyPlayingServiceReference = None
 		self.currentlyPlayingServiceOrGroup = None
 		self.currentlyPlayingService = None
+		self.originalPlayingServiceReference = None
 		self.currentServiceIsStreamRelay = False
 		self.skipServiceReferenceReset = False
 		self.RecordTimer = RecordTimer.RecordTimer()
@@ -106,20 +109,44 @@ class Navigation:
 	def restartService(self):
 		self.playService(self.currentlyPlayingServiceOrGroup, forceRestart=True)
 
-	def playService(self, ref, checkParentalControl=True, forceRestart=False, adjust=True):
+	def playService(self, ref, checkParentalControl=True, forceRestart=False, adjust=True, event=None):
+		# Some plugins send None as ref becasue want to shutdown enigma play system.
+		# So we have to stop current service if someone send None.
+		if ref is None:
+			self.stopService()
+			return 0
+		
+		InfoBarInstance = InfoBar.instance
+
 		oldref = self.currentlyPlayingServiceOrGroup
+		current_service_source = None
+		if InfoBarInstance:
+			current_service_source = InfoBarInstance.session.screen["CurrentService"]
+
+		if "%3a//" in ref.toString():
+			self.currentlyPlayingServiceReference = None
+			self.currentlyPlayingService = None
+			if current_service_source:
+				current_service_source.newService(False)
+
 		if ref and oldref and ref == oldref and not forceRestart:
 			print("[Navigation] ignore request to play already running service(1)")
 			return 1
 		print("[Navigation] playing ref", ref and ref.toString())
 
-		if exists("/proc/stb/lcd/symbol_signal"):
+		if exists("/proc/stb/lcd/symbol_signal") and hasattr(config.lcd, "mode"):
 			open("/proc/stb/lcd/symbol_signal", "w").write("1" if ref and "0:0:0:0:0:0:0:0:0" not in ref.toString() and config.lcd.mode.value else "0")
 
-		if ref is None:
-			self.stopService()
-			return 0
-		InfoBarInstance = InfoBar.instance
+		self.currentlyPlayingServiceReference = ref
+		self.currentlyPlayingServiceOrGroup = ref
+		self.originalPlayingServiceReference = ref
+
+		if InfoBarInstance and current_service_source:
+			current_service_source.serviceref = ref
+			current_service_source.newService(ref)
+			InfoBarInstance.session.screen["Event_Now"].updateSource(self.currentlyPlayingServiceReference)
+			InfoBarInstance.session.screen["Event_Next"].updateSource(self.currentlyPlayingServiceReference)
+			InfoBarInstance.serviceStarted()
 
 		if not checkParentalControl or parentalControl.isServicePlayable(ref, boundFunction(self.playService, checkParentalControl=False, forceRestart=forceRestart, adjust=adjust)):
 			if ref.flags & eServiceReference.isGroup:
@@ -139,6 +166,7 @@ class Navigation:
 							print("[Navigation] Failed to start: ", alternativeref.toString())
 							self.currentlyPlayingServiceReference = None
 							self.currentlyPlayingServiceOrGroup = None
+							self.originalPlayingServiceReference = None
 							if oldref and "://" in oldref.getPath():
 								print("[Navigation] Streaming was active -> try again")  # use timer to give the streamserver the time to deallocate the tuner
 								self.retryServicePlayTimer = eTimer()
@@ -160,7 +188,12 @@ class Navigation:
 					self.skipServiceReferenceReset = True
 				self.currentlyPlayingServiceReference = playref
 				playref = streamrelay.streamrelayChecker(playref)
+
+				for f in Navigation.playServiceExtensions:
+					playref, is_handled = f(self, playref, event, InfoBarInstance)
+
 				self.currentlyPlayingServiceOrGroup = ref
+
 				if InfoBarInstance and InfoBarInstance.servicelist.servicelist.setCurrent(ref, adjust):
 					self.currentlyPlayingServiceOrGroup = InfoBarInstance.servicelist.servicelist.getCurrent()
 				setPriorityFrontend = False
@@ -199,28 +232,24 @@ class Navigation:
 					self.retryServicePlayTimer = eTimer()
 					self.retryServicePlayTimer.callback.append(boundFunction(self.playService, ref, checkParentalControl, forceRestart, adjust))
 					self.retryServicePlayTimer.start(config.misc.softcam_streamrelay_delay.value, True)
-				else:
-					if self.pnav.playService(playref):
-						# print("[Navigation] Failed to start", playref)
+				elif not is_handled and self.pnav.playService(playref):
+						current_service_source.serviceref = None
 						self.currentlyPlayingServiceReference = None
+						self.originalPlayingServiceReference = None
 						self.currentlyPlayingServiceOrGroup = None
 						if oldref and "://" in oldref.getPath():
 							print("[Navigation] Streaming was active -> try again")  # use timer to give the streamserver the time to deallocate the tuner
 							self.retryServicePlayTimer = eTimer()
 							self.retryServicePlayTimer.callback.append(boundFunction(self.playService, ref, checkParentalControl, forceRestart, adjust))
 							self.retryServicePlayTimer.start(500, True)
-					else:  # self.pnav.playService was successful so force evStart here immediately because that doesn't happen for iptv services so the box appears frozen
-						self.currentlyPlayingServiceOrGroup = self.currentlyPlayingServiceReference
-						if InfoBarInstance:
-							InfoBarInstance.session.screen["CurrentService"].newService(self.currentlyPlayingServiceReference)
-							InfoBarInstance.session.screen["Event_Now"].updateSource(self.currentlyPlayingServiceReference)
-							InfoBarInstance.session.screen["Event_Next"].updateSource(self.currentlyPlayingServiceReference)
-							InfoBarInstance.serviceStarted()
 				self.skipServiceReferenceReset = False
 				if setPriorityFrontend:
 					setPreferredTuner(int(config.usage.frontend_priority.value))
 				if self.currentlyPlayingServiceReference and self.currentlyPlayingServiceReference.toString() in streamrelay.data:
 					self.currentServiceIsStreamRelay = True
+				if InfoBarInstance and "%3a//" in playref.toString() and not is_handled:
+					current_service_source.serviceref = None
+					InfoBarInstance.serviceStarted()
 				return 0
 		elif oldref and InfoBarInstance and InfoBarInstance.servicelist.servicelist.setCurrent(oldref, adjust):
 			self.currentlyPlayingServiceOrGroup = InfoBarInstance.servicelist.servicelist.getCurrent()
@@ -230,7 +259,12 @@ class Navigation:
 		return self.currentlyPlayingServiceReference
 
 	def getCurrentlyPlayingServiceOrGroup(self):
-		return self.currentlyPlayingServiceOrGroup
+		if not self.currentlyPlayingServiceOrGroup:
+			return None
+		return self.originalPlayingServiceReference or self.currentlyPlayingServiceOrGroup
+	
+	def getCurrentServiceReferenceOriginal(self):
+		return self.originalPlayingServiceReference
 
 	def getCurrentServiceRef(self):
 		curPlayService = self.getCurrentService()
@@ -245,6 +279,8 @@ class Navigation:
 			if ref.flags & eServiceReference.isGroup:
 				ref = getBestPlayableServiceReference(ref, eServiceReference(), simulate)
 			ref = streamrelay.streamrelayChecker(ref)
+			for f in Navigation.recordServiceExtensions:
+				ref = f(self, ref)
 			service = ref and self.pnav and self.pnav.recordService(ref, simulate)
 			if service is None:
 				print("[Navigation] record returned non-zero")
